+++ .vscode/settings.json
@@ -18,7 +18,8 @@
   "editor.tabSize": 2,
   "editor.insertSpaces": true,
   "[typescript]": {
-    "editor.formatOnSave": true
+    "editor.formatOnSave": true,
+    "editor.defaultFormatter": "xaver.clang-format"
   },
   "[javascript]": {
     "editor.formatOnSave": true
