+++ tfjs-core/src/tests.ts
@@ -40,9 +40,7 @@ import './kernel_registry_test';
 import './ops/arithmetic_test';
 import './ops/array_ops_test';
 import './ops/axis_util_test';
-import './ops/batchnorm2d_test';
-import './ops/batchnorm3d_test';
-import './ops/batchnorm4d_test';
+import './ops/batchnorm_test';
 import './ops/binary_ops_test';
 import './ops/boolean_mask_test';
 import './ops/broadcast_to_test';
